converted x-axis from coverage to efficacy
age_perc_vaxcov.py edits:
After running simulations, the ORs appear similar across
different vax efficacies even though the x-axis is the
only thing that should have changed. When plotted with coverage
in the x-axis, it appears that there is a slight increase in OR as
coverage (ie. efficacy) increases. Epidemic sizes don't vary
as much as expected.

Need to create a branch where we look at coverage in x-axis again
to see if the results are comparable.