removed inefficient percolation functions
There were three commented out functions that were
deleted. Two were percolation functions that used
only list searches and only dictionaries. These were
moved to bad_percolations.py for reference. The
flawed function that calculated incidence for the
general percolation scenarios was also removed
because it did not return a value. If needed, it can
be written to a cleaner form.