increase adult beta, sigma, gamma instead of child
Children tend to have larger epidemic peaks than adults all
of the time, but adults have relatively larger epidemic peaks
during severe seasons in the data. As adult attack rate increases,
total attack rate increases the most (due to the adult-centric
mixing matrix) and OR decreases, which matches with the data.
T-age, suscep, and recovery sims and viz scripts were updated
to focus on altering immunology of adults. Zipfile and file names
were changed to include "adult" in front of the name of the
simulation experiment.